requirements.txt
python-docx==1.2.0
pytesseract==0.3.13
pymupdf==1.26.3
langchain==0.3.27
langchain-community==0.3.27
langchain-ollama==0.3.6
pypdf==5.9.0
fastembed==0.7.1
faiss-cpu==1.11.0
streamlit==1.48.0